Prevented crash from faulty data
Chess.js library returns a location string with only a number and causes the player to be locked with the piece, not letting them move it at all. This code checks if the string is the correct length before allowing it through in hideMoves and showMoves to prevent the error from being thrown, as it runs just fine so long as that faulty object is excluded.